requirements.txt
# Orchestration & LLM
langchain
langgraph
langchain-community
langchain-huggingface
langchain-chroma

# RAG & Embeddings
sentence-transformers
chromadb

# Vision & Document Parsing
docling

# Voice & API
voxcpm 
torchcodec

# Data Processing
pandas
datasets<4
huggingface_hub

# Graphics
gradio

# Env and test
python-dotenv
pytest
PyYAML
